Accepts less than 1 output and work hours

+++ database/migrations/2015_12_07_192215_create_performances_table.php
@@ -23,9 +23,9 @@ class CreatePerformancesTable extends Migration
             $table->dateTime('date_start');
             $table->dateTime('date_end');
             $table->float('daily_work_hours');
-            $table->integer('output');
+            $table->float('output');
             $table->float('hours_worked');
-            $table->integer('output_error');
+            $table->float('output_error');
             $table->float('average_output');
             $table->timestamps();
             $table->softDeletes();
